Make everything modular by using header files
