+++ currency.h
@@ -1,0 +1,6 @@
+#ifndef  CURRENCY_H
+#define CURRENCY_H
+
+void conv_currency();
+
+#endif
