+++ mass.h
@@ -1,0 +1,5 @@
+#ifndef MASS_H
+#define MASS_H
+void conv_mass();
+
+#endif
