+++ temp.h
@@ -1,0 +1,1 @@
+void conv_temp();
