+++ unit.h
@@ -1,0 +1,6 @@
+#ifndef UNIT_H
+#define UNIT_H
+
+void ask_unit();
+
+#endif
